lab_mqtt_sistema_autenticacion/pyproject.toml
[project]
name = "lab-mqtt-sistema-autenticacion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "paho-mqtt>=2.1.0",
    "supabase>=2.28.3",
]
